+++ docs/src/components/anatomy/index.ts
@@ -1,0 +1,1 @@
+export { Anatomy } from './Anatomy';
