+++ docs/src/components/anatomy/types.ts
@@ -1,0 +1,5 @@
+import type { UseMasonryReturn } from 'react-virtual-masonry';
+
+// The library returns TanStack `VirtualItem`s but doesn't re-export the type;
+// derive it from the return shape so docs needn't depend on @tanstack directly.
+export type VirtualItem = UseMasonryReturn['items'][number];
