+++ docs/src/pages/anatomy/index.mdx
@@ -1,0 +1,7 @@
+# Anatomy [React Virtual Masonry, dissected]
+
+Drag the stage's right edge (or focus it and press ← / →) to resize the preview _container_ — lane count is driven by `@container` queries reading the container's own size, not the browser viewport. Scroll the stage and watch `MOUNTED` stay small while `ITEMS` grows as the feed auto-appends, with the mounted-index window shifting in the minimap.
+
+import { Anatomy } from '../../components/anatomy';
+
+<Anatomy />
