+++ docs/src/pages/docs/getting-started.mdx
@@ -106,4 +106,4 @@ See [`SSRConfig`](/docs/api/masonry#ssr) for the full set of SSR options.
 
 - Explore the [`<Masonry />`](/docs/api/masonry) props.
 - Need full control over the markup? Use the [`useMasonry`](/docs/api/use-masonry) hook directly.
-- See it in action on the [Demo](/demo) page.
+- See it in action on the [Anatomy](/anatomy) page.
